Adding some basic framework for QC validation plugins.
Not sure exactly where we can register these things, and I guess we can take in a config of (name,
fully-qualified-class-name, maybe-kwargs) pairs or something and build out the manager from that. I've had to create
stubs for the data model classes since I'm not sure how to test with django.